--- common/src/main/scala/hmda/model/census/County.scala
@@ -1,10 +1,0 @@
-package hmda.model.census
-
-case class County(
-    cbsaCode: String = "",
-    cbsaTitle: String = "",
-    countyName: String = "",
-    stateName: String = "",
-    stateCode: Int = 0,
-    countyCode: Int = 0
-)
