+++ common/src/main/scala/hmda/model/filing/EditDescriptionLookup.scala
@@ -1,10 +1,7 @@
 package hmda.model.filing
 
-import com.github.tototoshi.csv.CSVReader
 import com.typesafe.config.ConfigFactory
 import hmda.model.ResourceUtils._
-import hmda.model.census.County
-import scala.collection.immutable
 
 object EditDescriptionLookup {
 
@@ -15,8 +12,6 @@ object EditDescriptionLookup {
   val config = ConfigFactory.load()
   val editDescriptionFileName =
     config.getString("hmda.filing.edits.descriptions.filename")
-
-  val countyFileName = config.getString("hmda.county.countynames")
 
   val editDescriptionList = {
     val lines = fileLines(s"/$editDescriptionFileName")
@@ -29,27 +24,6 @@ object EditDescriptionLookup {
         val affectedDataFields = values(2).split(";").map(_.trim)
         EditDescription(editName, editDetails, affectedDataFields.toList)
       }
-  }
-
-  val countyList: Seq[County] = {
-    val reader = CSVReader.open(resource(s"/$countyFileName", "UTF-8"))
-    reader.toStream
-      .drop(1)
-      .map { field =>
-        County(field(0),
-               field(3),
-               field(7),
-               field(8),
-               field(9).trim().toInt,
-               field(10).trim().toInt)
-      }
-  }
-
-  val countiesMap =
-    countyList.map(e => (s"${e.stateCode}${e.countyCode}", e)).toMap
-
-  def lookupCounty(stateCode: Int, countyCode: Int): County = {
-    countiesMap.getOrElse(s"${stateCode}${countyCode}", County())
   }
 
   val editDescriptionMap: Map[String, EditDescription] =
